The Favorites, ProductDetail and ProductEdit index files were renamed to match the component name.
