+++ src/App.js
@@ -5,15 +5,15 @@ import LandingPage from './views/LandingPage/LandingPage';
 import Home from './views/Home/Home';
 import NavBar from "./components/NavBar/NavBar";
 import ContactUs from "./views/ContactUs";
-import ProductDetail from "./views/ProductDetail";
+import ProductDetail from "./views/ProductDetail/ProductDetail";
 import CreateProduct from "./views/CreateProduct";
 import UserProfile from "./views/UserProfile";
 import Payment from "./views/Payment";
 import NotFound from "./views/NotFound";
 import Footer from "./components/Footer/Footer";
 import AboutUs from "./views/AboutUs/AboutUs";
-import ProductEdit from "./views/ProductEdit";
-import Favorites from "./components/Favorites";
+import ProductEdit from "./views/ProductEdit/ProductEdit";
+import Favorites from "./components/Favorites/Favorites";
 import History from './components/History/History';
 import CookieCard from './components/CookieCard';
 // import StripeComponent from './components/StripeComponent';
