+++ src/components/Favorites/Favorites.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import './Favorites.css';
 import FavCard from "../FavCard";
